Disentangle `FragPart` from `autogen`
