+++ src/quemb/molbe/mbe.py
@@ -58,7 +58,7 @@ class BE(MixinLocalize):
     ----------
     mf : pyscf.scf.hf.SCF
         PySCF mean-field object.
-    fobj : quemb.molbe.autofrag.FragPart
+    fobj : quemb.molbe.fragment.FragPart
         Fragment object containing sites, centers, edges, and indices.
     eri_file : str
         Path to the file storing two-electron integrals.
